renames initialFormState as initialCreateFormState, moves to constants folder
